Add tests merge repos to verify the no WALs checks work
This adds some tests for #3908 as the new test API makes adding tests
for that issue much easier
